support variable length sha1 commit ids
The sha1 commit ids displayed by git may range in size and are not strictly 7 digits or 40 digits. In practice, the old default was something like 7, and that was enough to uniquely distinguish the commit. However, some repositories (Upstream Linux kernel) now use 12 as the default.

When we support intervals, use them to specify the full range to make the hint grab all available characters.