Match regexes for kubernetes api resources
  * Grabbed a list of resources with the cmd:
    kubectl api-resources -o name | xargs '-I{}' -n 1 echo -n '{}|'

  * Removed plural endings for resources
    (this matches the format of the clusters I interact with)

Note: I would like to add an optional "match an optional s" to handle
cases where the resources are pluralized, although I *think*
the way kubernetes names resources, they don't appear as plural
when doing `kubectl get all` which is what the hinting is designed
for. But I'm not super experienced with awk so I haven't figured
that one out yet.